refactor(desktop): isolate performance report formatting
Problem: Diagnostic enrichment exceeded the crash module file-size ratchet.

Root cause: The crash surface also owned all performance report text formatting.

Fix: Extract the formatter and keep its existing public export and report output.

Verification: Repository lint, frontend typecheck, and 99 crash-report assertions pass.
